Forgive the big commit
Fix compilation issues when starting oscar with empty functions or
unused variables

ID: 3613207 Fix escaping of names in Inbox
ID: 3611759 Restrict active demographics to auto complete list in inbox assign doc

Change prevention check PAP starting age to 21
Fixed display of prescriptions in eChart
Added immunization to ChildImmunization report for calculating bonuses
Restore archived status to drugs
Added document info such as type and upload date to list view of inbox

+++ src/main/java/org/oscarehr/common/dao/DocumentDao.java
@@ -168,18 +168,18 @@ public class DocumentDao extends AbstractDao<Document> {
 		return query.getResultList();
     }
 
-	public Long findMaxDocNo() {
+	public Integer findMaxDocNo() {
 		String sql = "select max(d.documentNo) from Document d";
 		Query query = entityManager.createQuery(sql);
 		List<Object> o = query.getResultList();
 		if (o.isEmpty()) {
-			return 0L;
+			return 0;
 		}
 		Object r = o.get(0);
 		if (r == null) {
-			return 0L;
+			return 0;
 		}
-		return ((Integer) r).longValue(); 
+		return (Integer) r; 
     }
 	
 	public Document getDocument(String documentNo) {
